requirements.txt
# Django et extensions
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-storages==1.14.2

# Documentation API
drf-yasg==1.21.7

# Channels pour WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Base de données
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Utilitaires
Pillow==10.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.5.0
pyjwt==2.8.0
celery==5.3.4
redis==5.0.1
boto3==1.28.63

# Sécurité
django-ratelimit==4.1.0

# Développement et tests
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.3.0
Faker==19.6.2
coverage==7.3.2
black==23.9.1
isort==5.12.0
flake8==6.1.0
